[REFAC: #34] - Isola views em pacotes.
close #34
