+++ app/core/views/__init__.py
@@ -1,0 +1,20 @@
+from .base import get_client_ip
+from .loan import (
+    loan_payment_list,
+    loan_payment_sum,
+    loan_retrieve,
+    loan_with_interest,
+    loans_lc,
+)
+from .payment import payment_lc, payment_retrieve
+
+__all__ = (
+    get_client_ip,
+    loans_lc,
+    loan_retrieve,
+    loan_with_interest,
+    loan_payment_sum,
+    loan_payment_list,
+    payment_lc,
+    payment_retrieve,
+)
