+++ app/core/views/loan.py
@@ -9,7 +9,7 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from app.core.models import Loan, Payment
+from app.core.models import Loan
 from app.core.permission import UserOnlyCanAccessOwnLoan, UserOnlyCanAccessOwnPayment
 from app.core.serializers import DebtLoanSerializer, LoanSerializer, PaymentSerializer, PaymentSumSerializer
 from app.core.services import extract_client_id, total_payment_for_the_loan
@@ -117,49 +117,8 @@ class LoanWithInterest(APIView):
         return Response(serialize.data)
 
 
-class PaymentLC(ListCreateAPIView):
-    serializer_class = PaymentSerializer
-    queryset = Payment.objects.all()
-
-    def get_queryset(self):
-        queryset = super().get_queryset()
-        return queryset.filter(loan__user=self.request.user)
-
-    @extend_schema(summary="Lista de pagamentos")
-    def get(self, request, *args, **kwargs):
-        """Retorna os pagamentos usuario. O usuario é
-        obtido pelo Token.
-        """
-        return super().get(request, *args, **kwargs)
-
-    @extend_schema(summary="Cria o pagamento para um emprestimo")
-    def post(self, request, *args, **kwargs):
-        """Cria o pagamento para emprestimo.O usuario é
-        obtido pelo Token.
-        """
-        return super().post(request, *args, **kwargs)
-
-
-class PaymentRetrieve(RetrieveAPIView):
-    serializer_class = PaymentSerializer
-    queryset = Payment.objects.all()
-    permission_classes = (UserOnlyCanAccessOwnPayment, IsAuthenticated)
-    lookup_field = "uuid"
-    lookup_url_kwarg = "id"
-
-    @extend_schema(summary="Recupera um pagamento")
-    def get(self, request, *args, **kwargs):
-        """Recupera um pagamento do usuario. O usuario é
-        obtido pelo Token.
-        """
-        return super().get(request, *args, **kwargs)
-
-
 loans_lc = LoansLC.as_view()
 loan_retrieve = LoanRetrieve.as_view()
 loan_payment_list = LoanPaymentList.as_view()
 loan_payment_sum = LoanPaymentSum.as_view()
 loan_with_interest = LoanWithInterest.as_view()
-
-payment_lc = PaymentLC.as_view()
-payment_retrieve = PaymentRetrieve.as_view()
